Add files via upload
client and server file is synchronous communication between client and server. Where a server can listen to a maximum of 4 number of clients. multiple clients can communicate with server. but server will be able to send respond to only on client
Tcpclients and server where clients can send data to the server and receive from the server.
blocking and non-blocking client
bufsize-receiving data with small buffersize in chunks.